+++ apps/ensadmin/src/components/indexing-status/index.ts
@@ -1,0 +1,1 @@
+export { IndexingStatus } from "./indexing-status";
