+++ apps/ensindexer/src/indexing-status/build-index-status.ts
@@ -20,7 +20,6 @@ import {
 } from "@ensnode/ensnode-sdk";
 import { prettifyError } from "zod/v4";
 
-import { parse } from "node:path";
 import config from "@/config";
 import ponderConfig from "@/ponder/config";
 import {
